LLM-CollabV1/requirements.txt
# API Libraries
requests==2.31.0
anthropic==0.2.6
openai==0.28.0
google-auth==2.21.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
google-api-python-client==2.93.0

# PyQt5 for UI
PyQt5==5.15.9
PyQt5-sip==12.11.1
PyQt5-Qt5==5.15.2
pygments==2.16.1  # Syntax highlighting
beautifulsoup4==4.12.2  # Parsing highlighted HTML

# Data Handling
jsonlib==1.6.1
simplejson==3.19.1

# Charting (used for visualizations)
PyQtChart==5.15.9

# Torch (potential usage in AI model execution)
torch==2.0.1

# Logging and Utilities
loguru==0.7.0
tqdm==4.65.0

# Optional but useful for managing environment variables
python-dotenv==1.0.0

# For secure HTTP (if required by any of the APIs)
cryptography==41.0.2
